cleaned up bugs

+++ RythmGame/Note.hpp
@@ -1,6 +1,5 @@
 #pragma once
 #include <SFML/System/Vector2.hpp>
-#include <SFML/Graphics/Color.hpp>
 
 class Note 
 {
@@ -11,13 +10,13 @@ public:
 		HOLD_START,
 		HOLD_END
 	};
-	Note(float hitTime, sf::Vector2f direction, int color, NoteType type);
-	Note(float hitTime, sf::Vector2f direction, int color, float holdNoteLength);
-	sf::Vector2f getPosition(long time);
+	Note(int hitTime, sf::Vector2f direction, int color, NoteType type);
+	Note(int hitTime, sf::Vector2f direction, int color, int holdNoteLength);
+	sf::Vector2f getPosition(int time);
 	void setEndPos(sf::Vector2f endPosIn);
 	int color;
-	float hitTime;
-	float holdNoteLength;
+	int hitTime;
+	int holdNoteLength;
 	NoteType noteType;
 private:
